fix AppConfig and validate netiv number

+++ random_progress/schemas.py
@@ -3,6 +3,6 @@ from dataclasses import dataclass
 
 @dataclass()
 class AppConfig:
-    min_number: int
-    max_number: int
-    sleep_time: int
+    min_number: float
+    max_number: float
+    sleep_time: float
